requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
pypdf==5.0.1
docx2txt==0.8
sentence-transformers==3.1.1
faiss-cpu==1.8.0
openai==1.45.1
python-dotenv==1.0.1
sqlalchemy==2.0.31
pymysql==1.1.0
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0